Added profit column in fixed deposits

+++ app/api/v1/schema/request/fixed_deposits.py
@@ -7,7 +7,7 @@ class FixedDepositsRequestSchema(TimeStampRequestSchema):
     rate_of_interest: str
     start_date: date
     end_date: date
-    maturity_amount: int | None = None
+    maturity_amount: float | None = None
     total_time: str
     remarks: str | None = None
     initial_investment: int
